# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plannify Driver openapi",
    "description": "API documentation for the Plannify Driver API",
    "contact": {
      "name": "contact@plannify.be"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api-dev.plannify.be/driver/v1"
    },
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/authentication/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Driver logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDriverResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/authentication/refresh": {
      "get": {
        "tags": [
          "authentication"
        ],
        "operationId": "refresh_token",
        "responses": {
          "200": {
            "description": "Driver auth refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDriverResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": []
      },
      "delete": {
        "tags": [
          "authentication"
        ],
        "operationId": "delete_refresh_token",
        "responses": {
          "200": {
            "description": "Driver auth deleted successfully"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/authentication/signup": {
      "post": {
        "tags": [
          "authentication"
        ],
        "operationId": "signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Driver signed up successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDriverResponse"
                }
              }
            }
          },
          "400": {
            "description": "Email domain denylisted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/authentication/token/verify-account": {
      "post": {
        "tags": [
          "authentication"
        ],
        "operationId": "verify_driver_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyDriverAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Driver account verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDriverResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDriverRequest": {
        "type": "object",
        "required": [
          "firstname",
          "lastname",
          "email",
          "password",
          "language"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "gender": {
            "type": [
              "string",
              "null"
            ]
          },
          "language": {
            "$ref": "#/components/schemas/DriverLanguage"
          },
          "lastname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "CreateDriverResponse": {
        "type": "object",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          }
        }
      },
...
      "DriverLanguage": {
        "type": "string",
        "enum": [
          "fr",
          "en"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "message",
          "error_code",
          "status"
        ],
        "properties": {
          "content": {},
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "LoginDriverRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "VerifyDriverAccountRequest": {
        "type": "object",
        "required": [
          "driver_id",
          "token"
        ],
        "properties": {
          "driver_id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
    }
  }
}